+++ plugins/gui/include/gui/module_widget/module_widget.h
@@ -53,6 +53,7 @@ namespace hal
 
         virtual void setupToolbar(Toolbar* Toolbar) override;
         virtual QList<QShortcut*> createShortcuts() override;
+        void openModuleInView(u32 moduleId);
 
     public Q_SLOTS:
         void toggleSearchbar();